+++ mellon/__init__.py
@@ -19,7 +19,7 @@ from . import _conditional as conditional
 from . import _derivatives as derivatives
 from . import validation
 
-__version__ = "1.4.3"
+__version__ = "1.5.0"
 
 __all__ = [
     "DensityEstimator",