+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - remove `numpy` as direct dependency
  - bugfix DimensionalityEstimator dimensionality initialization
  - implement 'fixed' gaussian proces type to allow more inducing points than datapoints
+ - implement `copy()` method for `Predictor` class
 
 # v1.4.3